+++ Arbiter.App/ViewModels/Dialogs/DialogManagerViewModel.cs
@@ -2,8 +2,11 @@
 using System.Collections.Concurrent;
 using System.Collections.ObjectModel;
 using System.Collections.Specialized;
+using System.Threading.Tasks;
+using Arbiter.App.Extensions;
 using Arbiter.App.ViewModels.Client;
 using Arbiter.Net.Proxy;
+using Avalonia;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 using Microsoft.Extensions.DependencyInjection;
@@ -27,8 +30,6 @@ public partial class DialogManagerViewModel : ViewModelBase
 
     [ObservableProperty] private bool _shouldSync = true;
 
-    public ObservableCollection<ClientViewModel> Clients => _clientManager.Clients;
-
     public DialogManagerViewModel(ILogger<DialogManagerViewModel> logger, IServiceProvider serviceProvider)
     {
         _logger = logger;
@@ -44,12 +45,6 @@ public partial class DialogManagerViewModel : ViewModelBase
 
     private void OnClientSelected(ClientViewModel? client)
     {
-        if (client is null || SelectedClient is not null)
-        {
-            return;
-        }
-
-        // Automatically select the client if none is selected
         SelectedClient = client;
     }
 
@@ -135,6 +130,18 @@ public partial class DialogManagerViewModel : ViewModelBase
         dialog.RequestNext -= OnDialogNavigateNext;
         dialog.RequestTop -= OnDialogNavigateTop;
         dialog.RequestClose -= OnDialogClose;
+    }
+
+    [RelayCommand]
+    private async Task CopyDialogTextToClipboard()
+    {
+        var clipboard = Application.Current?.TryGetClipboard();
+        if (clipboard is null || ActiveDialog is null)
+        {
+            return;
+        }
+
+        await clipboard.SetTextAsync(ActiveDialog.Content);
     }
 
     [RelayCommand]