+++ Arbiter.App/ViewModels/Entities/EntityManagerViewModel.PacketFilters.cs
@@ -1,6 +1,8 @@
 ﻿using System;
+using System.Collections.Concurrent;
 using Arbiter.App.Models;
 using Arbiter.Net;
+using Arbiter.Net.Client.Messages;
 using Arbiter.Net.Filters;
 using Arbiter.Net.Proxy;
 using Arbiter.Net.Server.Messages;
@@ -9,65 +11,108 @@ using Arbiter.Net.Types;
 
 namespace Arbiter.App.ViewModels.Entities;
 
-public partial class EntityListViewModel
+public partial class EntityManagerViewModel
 {
+    private const string FilterPrefix = nameof(EntityManagerViewModel);
+    private static readonly TimeSpan InteractRequestTimeout = TimeSpan.FromSeconds(2);
+
+    // Used to track interaction queries for a client to map back to the correct entity
+    private readonly ConcurrentDictionary<int, ConcurrentQueue<(uint, DateTime)>> _interactRequests = [];
+
     private void AddPacketFilters(ProxyServer proxyServer)
     {
         proxyServer.AddFilter(new ServerMessageFilter<ServerAddEntityMessage>(OnAddEntityMessage)
         {
-            Name = "EntityView_AddEntityFilter",
+            Name = $"{FilterPrefix}_ServerAddEntity",
             Priority = int.MaxValue - 10
         });
 
         proxyServer.AddFilter(new ServerMessageFilter<ServerShowUserMessage>(OnShowUserMessage)
         {
-            Name = "EntityView_ShowUserFilter",
+            Name = $"{FilterPrefix}_ServerShowUser",
             Priority = int.MaxValue - 10
         });
-        
+
         proxyServer.AddFilter(new ServerMessageFilter<ServerUserProfileMessage>(OnUserProfileMessage)
         {
-            Name = "EntityView_UserProfileFilter",
+            Name = $"{FilterPrefix}_ServerUserProfile",
             Priority = int.MaxValue - 10
         });
 
         proxyServer.AddFilter(new ServerMessageFilter<ServerPublicMessageMessage>(OnPublicMessage)
         {
-            Name = "EntityView_PublicMessageFilter",
+            Name = $"{FilterPrefix}_ServerPublicMessage",
             Priority = int.MaxValue - 10
         });
 
         proxyServer.AddFilter(new ServerMessageFilter<ServerShowDialogMessage>(OnShowDialogMessage)
         {
-            Name = "EntityView_ShowDialogFilter",
+            Name = $"{FilterPrefix}_ServerShowDialog",
             Priority = int.MaxValue - 10
         });
 
         proxyServer.AddFilter(new ServerMessageFilter<ServerShowDialogMenuMessage>(OnShowDialogMenuMessage)
         {
-            Name = "EntityView_ShowDialogMenuFilter",
+            Name = $"{FilterPrefix}_ServerShowDialogMenu",
             Priority = int.MaxValue - 10
         });
 
         proxyServer.AddFilter(new ServerMessageFilter<ServerEntityWalkMessage>(OnEntityWalkMessage)
         {
-            Name = "EntityView_EntityWalkFilter",
+            Name = $"{FilterPrefix}_ServerEntityWalk",
             Priority = int.MaxValue - 10
         });
 
         proxyServer.AddFilter(new ServerMessageFilter<ServerWalkResponseMessage>(OnSelfWalkMessage)
         {
-            Name = "EntityView_SelfWalkFilter",
+            Name = $"{FilterPrefix}_ServerWalkResponse",
+            Priority = int.MaxValue - 10
+        });
+
+        proxyServer.AddFilter(new ServerMessageFilter<ServerMapLocationMessage>(OnServerMapLocationMessage)
+        {
+            Name = $"{FilterPrefix}_ServerMapLocation",
             Priority = int.MaxValue - 10
         });
 
         proxyServer.AddFilter(new ServerMessageFilter<ServerRemoveEntityMessage>(OnRemoveEntityMessage)
         {
-            Name = "EntityView_RemoveEntityFilter",
+            Name = $"{FilterPrefix}_ServerRemoveEntity",
             Priority = int.MaxValue - 10
         });
+
+        proxyServer.AddFilter(
+            new ClientMessageFilter<ClientInteractMessage>(HandleInteractMessage)
+            {
+                Name = $"{FilterPrefix}_ClientInteract",
+                Priority = int.MaxValue - 10
+            });
+
+        proxyServer.AddFilter(
+            new ServerMessageFilter<ServerWorldMessageMessage>(HandleInteractResponseMessage)
+            {
+                Name = $"{FilterPrefix}_ServerWorldMessage",
+                Priority = int.MaxValue - 10
+            });
+
+        proxyServer.ClientConnected += OnClientConnected;
+        proxyServer.ClientDisconnected += OnClientDisconnected;
     }
 
+    private void OnClientConnected(object? sender, ProxyConnectionEventArgs e)
+        => _interactRequests.AddOrUpdate(e.Connection.Id, [], (_, _) => []);
+
+    private void OnClientDisconnected(object? sender, ProxyConnectionEventArgs e)
+        => _interactRequests.TryRemove(e.Connection.Id, out _);
+
+    private void QueueInteractionRequest(int connectionId, uint entityId)
+    {
+        if (_interactRequests.TryGetValue(connectionId, out var queue))
+        {
+            queue.Enqueue((entityId, DateTime.Now));
+        }
+    }
+    
     private NetworkPacket OnAddEntityMessage(ProxyConnection connection, ServerAddEntityMessage message,
         object? parameter, NetworkMessageFilterResult<ServerAddEntityMessage> result)
     {
@@ -134,7 +179,7 @@ public partial class EntityListViewModel
         // Do not alter the packet
         return result.Passthrough();
     }
-    
+
     private NetworkPacket OnUserProfileMessage(ProxyConnection connection, ServerUserProfileMessage message,
         object? parameter, NetworkMessageFilterResult<ServerUserProfileMessage> result)
     {
@@ -155,7 +200,7 @@ public partial class EntityListViewModel
         // Do not alter the packet
         return result.Passthrough();
     }
-    
+
     private NetworkPacket OnPublicMessage(ProxyConnection connection, ServerPublicMessageMessage message,
         object? parameter, NetworkMessageFilterResult<ServerPublicMessageMessage> result)
     {
@@ -164,23 +209,24 @@ public partial class EntityListViewModel
         {
             return result.Passthrough();
         }
-        
+
         // Try to get the player so we can get map context
         _playerService.TryGetState(connection.Id, out var player);
-        
+
         // Assume it might be a player ghost
         var entityFlags = EntityFlags.Player;
         var entitySprite = (ushort)BodySprite.MaleGhost;
-        
+
         // Try to get the existing entity, this should rule out monsters/npcs that talk
         if (_entityStore.TryGetEntity(message.SenderId, out var existing))
         {
             entityFlags = existing.Flags;
             entitySprite = existing.Sprite ?? entitySprite;
         }
-        
+
         // The name should come before the symbol, so split on the first symbol (chat or shout)
-        var senderName = message.Message.Split(':', '!', StringSplitOptions.RemoveEmptyEntries | StringSplitOptions.TrimEntries)[0];
+        var senderName =
+            message.Message.Split(':', '!', StringSplitOptions.RemoveEmptyEntries | StringSplitOptions.TrimEntries)[0];
         var entity = new GameEntity
         {
             Flags = entityFlags,
@@ -287,29 +333,32 @@ public partial class EntityListViewModel
     private NetworkPacket OnEntityWalkMessage(ProxyConnection connection, ServerEntityWalkMessage message,
         object? parameter, NetworkMessageFilterResult<ServerEntityWalkMessage> result)
     {
-        var existing = _entityStore.TryGetEntity(message.EntityId, out var entity);
+        var existing = _entityStore.TryGetEntity(message.EntityId, out _);
         if (!existing)
         {
             return result.Passthrough();
         }
 
-        var newEntity = entity with
+        var newX = message.Direction switch
         {
-            X = message.Direction switch
-            {
-                WorldDirection.Left => Math.Max(0, message.OriginX - 1),
-                WorldDirection.Right => message.OriginX + 1,
-                _ => message.OriginX
-            },
-            Y = message.Direction switch
-            {
-                WorldDirection.Up => Math.Max(0, message.OriginY - 1),
-                WorldDirection.Down => message.OriginY + 1,
-                _ => message.OriginY
-            }
+            WorldDirection.Left => Math.Max(0, message.OriginX - 1),
+            WorldDirection.Right => message.OriginX + 1,
+            _ => message.OriginX
         };
 
-        _entityStore.AddOrUpdateEntity(newEntity, out _);
+        var newY = message.Direction switch
+        {
+            WorldDirection.Up => Math.Max(0, message.OriginY - 1),
+            WorldDirection.Down => message.OriginY + 1,
+            _ => message.OriginY
+        };
+
+        _entityStore.TrySetEntityLocation(message.EntityId, newX, newY);
+
+        if (SelectedClient is not null && FilterMode is not EntityFilterMode.All)
+        {
+            _filterDebouncer.Execute(RefreshFilterPreservingSelection);
+        }
 
         // Do not alter the packet
         return result.Passthrough();
@@ -348,6 +397,103 @@ public partial class EntityListViewModel
             }
         };
         _entityStore.AddOrUpdateEntity(newEntity, out _);
+
+        if (SelectedClient is not null)
+        {
+            _filterDebouncer.Execute(RefreshFilterPreservingSelection);
+        }
+
+        // Do not alter the packet
+        return result.Passthrough();
+    }
+
+    private NetworkPacket OnServerMapLocationMessage(ProxyConnection connection, ServerMapLocationMessage message,
+        object? parameter, NetworkMessageFilterResult<ServerMapLocationMessage> result)
+    {
+        if (SelectedClient is not null && FilterMode is not EntityFilterMode.All)
+        {
+            _filterDebouncer.Execute(RefreshFilterPreservingSelection);
+        }
+
+        // Do not alter the packet
+        return result.Passthrough();
+    }
+
+    private NetworkPacket HandleInteractMessage(ProxyConnection connection, ClientInteractMessage message,
+        object? parameter, NetworkMessageFilterResult<ClientInteractMessage> result)
+    {
+        // Ensure there is a valid entity interaction request
+        if (message.InteractionType != InteractionType.Entity || !message.TargetId.HasValue)
+        {
+            return result.Passthrough();
+        }
+
+        // Determine which entity type we are interacting with
+        var entityFlags = EntityFlags.None;
+        if (_entityStore.TryGetEntity(message.TargetId.Value, out var entity))
+        {
+            entityFlags = entity.Flags;
+        }
+
+        // If it's a player, npc, or item we can ignore it
+        if (entityFlags.HasFlag(EntityFlags.Player) || entityFlags.HasFlag(EntityFlags.Mundane) ||
+            entityFlags.HasFlag(EntityFlags.Item))
+        {
+            return result.Passthrough();
+        }
+
+        QueueInteractionRequest(connection.Id, message.TargetId.Value);
+
+        // Do not alter the packet
+        return result.Passthrough();
+    }
+
+    private NetworkPacket HandleInteractResponseMessage(ProxyConnection connection, ServerWorldMessageMessage message,
+        object? parameter, NetworkMessageFilterResult<ServerWorldMessageMessage> result)
+    {
+        var trimmedMessage = message.Message.Trim();
+
+        // If the message is empty it's not a monster name
+        if (string.IsNullOrWhiteSpace(trimmedMessage))
+        {
+            return result.Passthrough();
+        }
+
+        // If the message ends with a period, question mark, or exclamation mark then it's not a monster name
+        if (trimmedMessage.EndsWith('.') || trimmedMessage.EndsWith('!') || trimmedMessage.EndsWith('?'))
+        {
+            return result.Passthrough();
+        }
+
+        // We can safely ignore any message that contains any of the following characters
+        foreach (var c in trimmedMessage)
+        {
+            if (c is '[' or ']' or '(' or ')' or '<' or '>' or ':' or '!' or '?' or '"' or ',')
+            {
+                return result.Passthrough();
+            }
+        }
+
+        // If there is no queue or it is empty then not pending interactions
+        if (!_interactRequests.TryGetValue(connection.Id, out var queue) || queue.IsEmpty)
+        {
+            return result.Passthrough();
+        }
+
+        // Look for a request that has not timed out
+        var now = DateTime.Now;
+        while (queue.TryDequeue(out var pair))
+        {
+            var (entityId, timestamp) = pair;
+            var elapsed = now - timestamp;
+
+            if (elapsed > InteractRequestTimeout)
+            {
+                continue;
+            }
+
+            _entityStore.TrySetEntityName(entityId, trimmedMessage);
+        }
 
         // Do not alter the packet
         return result.Passthrough();