+++ Arbiter.App/ViewModels/Proxy/ProxyViewModel.Filters.cs
@@ -5,6 +5,8 @@ namespace Arbiter.App.ViewModels.Proxy;
 
 public partial class ProxyViewModel
 {
+    private const string FilterPrefix = nameof(ProxyViewModel);
+    
     // Almost highest, but allows for other filters to run first
     private const int DebugFilterPriority = int.MaxValue - 100;
 
@@ -21,6 +23,7 @@ public partial class ProxyViewModel
         AddDebugEffectsFilters(settings);
         AddDebugMapFilters(settings);
         AddDebugMessageFilters(settings, filters);
+        AddModMenuFilters(settings);
     }
 
     public void RemoveDebugFilters()
@@ -31,5 +34,6 @@ public partial class ProxyViewModel
         RemoveDebugEffectsFilters();
         RemoveDebugMapFilters();
         RemoveDebugMessageFilters();
+        RemoveModMenuFilters();
     }
 }